# Consolidates buffers and generally improves string decoding