--- a/zipkin/src/main/java/zipkin2/internal/Proto3Fields.java
+++ b/zipkin/src/main/java/zipkin2/internal/Proto3Fields.java
@@ -124,18 +124,9 @@ final class Proto3Fields {
      * is returned when the length prefix is zero.
      */
     final T readLengthPrefixAndValue(UnsafeBuffer b) {
-      int length = guardLength(b);
+      int length = b.readVarint32();
       if (length == 0) return null;
       return readValue(b, length);
-    }
-
-    final int guardLength(UnsafeBuffer buffer) {
-      int length = buffer.readVarint32();
-      if (length > buffer.remaining()) {
-        throw new IllegalArgumentException(
-          "Truncated: length " + length + " > bytes remaining " + buffer.remaining());
-      }
-      return length;
     }
 
     abstract int sizeOfValue(T value);