This consolidates buffers used for string decoding in binary protocols and uses them in thrift and proto. It also special cases to get better parsing performance than 3rd party codec libraries

before
```
CodecBenchmarks.readChineseSpan_PROTO3:readChineseSpan_PROTO3·p0.99       sample              3.036             us/op
CodecBenchmarks.readChineseSpan_PROTO3:·gc.alloc.rate.norm                sample      15   2328.293 ±   0.024    B/op
CodecBenchmarks.readChineseSpan_THRIFT:readChineseSpan_THRIFT·p0.99       sample              8.192             us/op
CodecBenchmarks.readChineseSpan_THRIFT:·gc.alloc.rate.norm                sample      15   6464.752 ±   0.078    B/op
CodecBenchmarks.readClientSpan_PROTO3:readClientSpan_PROTO3·p0.99         sample              2.248             us/op
CodecBenchmarks.readClientSpan_PROTO3:·gc.alloc.rate.norm                 sample      15   2104.260 ±   0.023    B/op
CodecBenchmarks.readClientSpan_THRIFT:readClientSpan_THRIFT·p0.99         sample             12.176             us/op
CodecBenchmarks.readClientSpan_THRIFT:·gc.alloc.rate.norm                 sample      15   6184.692 ±  12.531    B/op
```

after
```
CodecBenchmarks.readChineseSpan_PROTO3:readChineseSpan_PROTO3·p0.99       sample              2.316             us/op
CodecBenchmarks.readChineseSpan_PROTO3:·gc.alloc.rate.norm                sample      15   2032.257 ±   0.025    B/op
CodecBenchmarks.readChineseSpan_THRIFT:readChineseSpan_THRIFT·p0.99       sample              7.511             us/op
CodecBenchmarks.readChineseSpan_THRIFT:·gc.alloc.rate.norm                sample      15   5461.857 ±  41.728    B/op
CodecBenchmarks.readClientSpan_PROTO3:readClientSpan_PROTO3·p0.99         sample              1.936             us/op
CodecBenchmarks.readClientSpan_PROTO3:·gc.alloc.rate.norm                 sample      15   1744.273 ±   0.015    B/op
CodecBenchmarks.readClientSpan_THRIFT:readClientSpan_THRIFT·p0.99         sample              7.472             us/op
CodecBenchmarks.readClientSpan_THRIFT:·gc.alloc.rate.norm                 sample      15   4920.538 ± 187.788    B/op
```