@adriancole 2019-05-12
TODO unit tests