@adriancole 2019-05-12
Here's the comparative benchmark results now:

```
ProtoCodecBenchmarks.bytes_protobufDecoder:bytes_protobufDecoder·p0.99       sample            1935.360                us/op
ProtoCodecBenchmarks.bytes_protobufDecoder:·gc.alloc.rate.norm               sample     15  2023207.729 ±     32.000    B/op
ProtoCodecBenchmarks.bytes_wireDecoder:bytes_wireDecoder·p0.99               sample            2389.402                us/op
ProtoCodecBenchmarks.bytes_wireDecoder:·gc.alloc.rate.norm                   sample     15  2435412.109 ±     55.776    B/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:bytes_zipkinDecoder·p0.99           sample            1724.416                us/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.alloc.rate.norm                 sample     15  1735155.292 ±     29.521    B/op
```