@adriancole 2019-05-12
@anuraaga I decided to pick back up the making it possible to use ByteBuffer thing. Rationale was that it is not going as well as I hoped to transfer into a different array per span.

The work I have in progress builds on this.. slowing a little to make buffer work.  still.. it is as fast as any library we've used yet.